requirements.txt
tensorflow==2.9
keras==2.9
numpy==1.23.*
pandas==1.4.*
seaborn
matplotlib==3.5.3
scipy
pyyaml
argparse